+++ src/spec/SoundSpec.ts
@@ -144,6 +144,16 @@ describe('Sound resource', () => {
          
          expect(audioInstance.volume).toBe(0.5);
       });
+      
+      it('should set volume with argument sent to play', () => {
+         var audioInstance = new MockAudioInstance();
+         
+         spyOn(audioMock, 'createInstance').and.returnValue(audioInstance);
+
+         sut.play(0.5);
+         
+         expect(audioInstance.volume).toBe(0.5);
+      });
 
       it('should play once and then finish', (done) => {
          sut.play().then(() => {