+++ CHANGELOG.md
@@ -10,8 +10,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Breaking Changes
 +- Added `pos` property of type: `Vector` to `PointerEvent` using `x` and `y` arguments from constructor. `x` and `y` are still public properties of `PointerEvent`, but are now marked as @deprecated ([#612](https://github.com/excaliburjs/Excalibur/issues/612))
 ### Added
+- Added optional volume argument to `Sound.play(volume?: number)`, which will play the Audio file
+at anywhere from mute (`volume` is 0.0) to full volume (`volume` is 1.0). ([#801](https://github.com/excaliburjs/Excalibur/issues/801))
 - Added another DisplayMode option: `DisplayMode.Position`. When this is selected as the displayMode type, the user must specify a new `position` option
 - Added a new Interface `AbsolutePosition` which can described the `position` option. A `string` can also describe `position`
+
 ### Changed
 - Edge builds have more descriptive versions now containing build number and Git commit hash (e.g. `0.10.0-alpha.105#commit`) ([#777](https://github.com/excaliburjs/Excalibur/issues/777))
 ### Deprecated